Convert file names to integer in stmThreadList
File names in the stmThreadList function are now converted to integers. This change ensures that the function returns an array of integer values rather than strings, making the file name data easier to manage and process.

+++ Src/Process/Kube/Prime/PrimeHandler.php
@@ -21,7 +21,7 @@ trait PrimeHandler
     public static function stmThreadList(): array
     {
         $files = glob(static::$STM_THREADS_PATH . '/*.json');
-        foreach ($files as $key => $path) $files[$key] = basename($path, '.json');
+        foreach ($files as $key => $path) $files[$key] = (int) basename($path, '.json');
         return $files;
     }
